execute error with this class
when i use hive-jdbc to get data from hive, i got this exception
org.apache.commons.beanutils.ConversionException: Cannot assign value of type 'java.math.BigDecimal' to property 'valuess' of type 'java.math.BigInteger'
because the data type is decimal
so i change the type to solve it

+++ jdbc/src/java/org/apache/hive/jdbc/JdbcColumn.java
@@ -18,7 +18,7 @@
 
 package org.apache.hive.jdbc;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 import java.sql.Date;
 import java.sql.SQLException;
 import java.sql.Timestamp;
@@ -98,7 +98,7 @@ public class JdbcColumn {
       case Types.TIMESTAMP_WITH_TIMEZONE:
         return TimestampTZ.class.getName();
       case Types.DECIMAL:
-        return BigInteger.class.getName();
+        return BigDecimal.class.getName();
       case Types.BINARY:
         return byte[].class.getName();
       case Types.OTHER:
